+++ src/components/organisms/Posts.tsx
@@ -1,10 +1,8 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import styled from 'styled-components'
 import { Post } from './Post'
-import { getPosts } from '../../lib/api/tweet'
 import { Loading } from '../pages/Loading'
 import { Pagination } from '@mui/material'
-import { useLocation, useNavigate } from 'react-router-dom'
 
 type Post = {
   id: number
@@ -20,77 +18,21 @@ type Post = {
   likes: number
 }
 
-// TODO: リツイート、いいね機能、アバター画像後に削除する
-const initialPost: Post = {
-  id: 1,
-  user: {
-    id: 'u1',
-    name: 'aliiiiii1',
-    nickname: 'Alice',
-    avatarUrl: '/path/to/avatar1.png',
-  },
-  tweet: 'This is a sample tweet from Alice.',
-  imageUrl: 'https://source.unsplash.com/random',
-  retweets: 5,
-  likes: 20,
+type Props = {
+  posts: Post[]
+  loading: boolean
+  currentPage: number
+  totalPages: number
+  handleChangePage: (_event: React.ChangeEvent<unknown>, page: number) => void
 }
 
-export const Posts: React.FC = () => {
-  const [posts, setPosts] = useState<Post[]>([])
-  const [loading, setLoading] = useState<boolean>(false)
-  const [currentPage, setCurrentPage] = React.useState(1)
-  const [totalPages, setTotalPages] = React.useState(1)
-
-  const navigate = useNavigate()
-  const location = useLocation()
-
-  const handleChangePage = (
-    _event: React.ChangeEvent<unknown>,
-    page: number
-  ) => {
-    setCurrentPage(page)
-    navigate({ ...location, search: `?page=${page}` })
-    handleGetPosts(page)
-    window.scrollTo(0, 0)
-  }
-
-  const handleGetPosts = (page: number) => {
-    setLoading(true)
-
-    getPosts(page)
-      .then((res) => {
-        if (res && res.data) {
-          setTotalPages((res.data.pagination.totalPages as number) || 1)
-          const tmpPosts: Post[] = (res.data.tweets as Post[]).map((tweet) => {
-            return {
-              ...initialPost,
-              id: tweet.id,
-              user: tweet.user,
-              tweet: tweet.tweet,
-              imageUrl: tweet.imageUrl,
-              retweets: tweet.retweets,
-              likes: tweet.likes,
-            }
-          })
-          setPosts(tmpPosts)
-        }
-      })
-      .catch((err) => {
-        console.error(err)
-      })
-      .finally(() => {
-        setLoading(false)
-      })
-  }
-
-  useEffect(() => {
-    //  クエリパラメータからpageを取得する
-    const queryParams = new URLSearchParams(location.search)
-    const page = Number(queryParams.get('page')) || 1
-    setCurrentPage(page)
-    handleGetPosts(page)
-  }, [location.search])
-
+export const Posts: React.FC<Props> = ({
+  posts,
+  loading,
+  currentPage,
+  totalPages,
+  handleChangePage,
+}) => {
   if (loading) {
     return <Loading />
   }
