+++ src/components/templates/TimeLine.tsx
@@ -1,16 +1,107 @@
-import React from 'react'
+import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
 import { TweetBox } from '../organisms/TweetBox'
 import { Posts } from '../organisms/Posts'
+import { getPosts } from '../../lib/api/tweet'
+import { useNavigate } from 'react-router-dom'
+
+type Post = {
+  id: number
+  user: {
+    id: string
+    name: string
+    nickname: string
+    avatarUrl: string
+  }
+  tweet: string
+  imageUrl: string
+  retweets: number
+  likes: number
+}
+
+// TODO: リツイート、いいね機能、アバター画像後に削除する
+const initialPost: Post = {
+  id: 1,
+  user: {
+    id: 'u1',
+    name: 'aliiiiii1',
+    nickname: 'Alice',
+    avatarUrl: '/path/to/avatar1.png',
+  },
+  tweet: 'This is a sample tweet from Alice.',
+  imageUrl: 'https://source.unsplash.com/random',
+  retweets: 5,
+  likes: 20,
+}
 
 export const TimeLine: React.FC = () => {
+  const [posts, setPosts] = useState<Post[]>([])
+  const [loading, setLoading] = useState<boolean>(false)
+  const [currentPage, setCurrentPage] = React.useState(1)
+  const [totalPages, setTotalPages] = React.useState(1)
+
+  const navigate = useNavigate()
+
+  const handleChangePage = (
+    _event: React.ChangeEvent<unknown>,
+    page: number
+  ) => {
+    setCurrentPage(page)
+    navigate({ ...location, search: `?page=${page}` })
+    handleGetPosts(page)
+  }
+
+  const handleGetPosts = (page: number) => {
+    setLoading(true)
+
+    getPosts(page)
+      .then((res) => {
+        if (res && res.data) {
+          setTotalPages((res.data.pagination.totalPages as number) || 1)
+          const resPosts: Post[] = (res.data.tweets as Post[]).map((tweet) => {
+            return {
+              ...initialPost,
+              id: tweet.id,
+              user: tweet.user,
+              tweet: tweet.tweet,
+              imageUrl: tweet.imageUrl,
+              retweets: tweet.retweets,
+              likes: tweet.likes,
+            }
+          })
+          setPosts(resPosts)
+        }
+      })
+      .catch((err) => {
+        console.error(err)
+      })
+      .finally(() => {
+        setLoading(false)
+      })
+  }
+
+  useEffect(() => {
+    //  クエリパラメータからpageを取得する
+    const queryParams = new URLSearchParams(location.search)
+    const page = Number(queryParams.get('page')) || 1
+
+    setCurrentPage(page)
+    handleGetPosts(page)
+  }, [location.search])
+
   return (
     <StyledTimeLine>
       <div className="timeLineHeader">
         <h2>ホーム</h2>
-        <TweetBox />
+        <TweetBox handleGetPosts={handleGetPosts} />
       </div>
-      <Posts />
+      <Posts
+        posts={posts}
+        loading={loading}
+        currentPage={currentPage}
+        totalPages={totalPages}
+        handleChangePage={handleChangePage}
+      />
     </StyledTimeLine>
   )
 }
